+++ src/lib.rs
@@ -19,6 +19,6 @@ mod stroker;
 mod vid;
 
 pub use geom::{Pt, Transform};
-pub use path::{FillRule, Path, PathOp};
+pub use path::{FillRule, Path2D, PathOp};
 pub use plotter::Plotter;
 pub use stroker::JoinStyle;